+++ libs/angular-remote-components/src/lib/angular-remote-components.module.ts
@@ -1,12 +1,17 @@
 import { NgModule } from '@angular/core'
 import { CommonModule } from '@angular/common'
 import { SlotComponent } from './components/slot/slot.component'
-import { AppConfigService } from '@onecx/angular-accelerator'
+import { SLOT_SERVICE, SlotService } from './services/slot.service'
 
 @NgModule({
   imports: [CommonModule],
   declarations: [SlotComponent],
   exports: [SlotComponent],
-  providers: [AppConfigService],
+  providers: [
+    {
+      provide: SLOT_SERVICE,
+      useExisting: SlotService,
+    },
+  ],
 })
 export class AngularRemoteComponentsModule {}