+++ libs/angular-testing/src/lib/harnesses/primeng/p-dropdown.harness.ts
@@ -3,15 +3,26 @@ import { ListItemHarness } from '../list-item.harness'
 
 export interface PDropdownHarnessFilters extends BaseHarnessFilters {
   id?: string
+  inputId?: string
 }
 
 export class PDropdownHarness extends ContentContainerComponentHarness {
   static hostSelector = 'p-dropdown'
 
   static with(options: PDropdownHarnessFilters): HarnessPredicate<PDropdownHarness> {
-    return new HarnessPredicate(PDropdownHarness, options).addOption('id', options.id, (harness, id) =>
-      HarnessPredicate.stringMatches(harness.getId(), id)
-    )
+    return new HarnessPredicate(PDropdownHarness, options)
+      .addOption('id', options.id, (harness, id) => HarnessPredicate.stringMatches(harness.getId(), id))
+      .addOption('inputId', options.inputId, (harness, inputId) =>
+        HarnessPredicate.stringMatches(harness.getInputId(), inputId)
+      )
+  }
+
+  async getInputId(): Promise<string | null> {
+    return await (await this.host()).getAttribute('inputId')
+  }
+
+  async getAriaLabel(): Promise<string | null | undefined> {
+    return (await this.locatorForOptional('span.p-placeholder')())?.getAttribute('aria-label')
   }
 
   async getId(): Promise<string | null> {
@@ -20,6 +31,10 @@ export class PDropdownHarness extends ContentContainerComponentHarness {
 
   async getDefaultText() {
     return (await this.locatorForOptional('span.p-placeholder')())?.text()
+  }
+
+  async getSelectedText() {
+    return (await this.locatorForOptional('span.p-dropdown-label')())?.text()
   }
 
   async isOpen(): Promise<boolean> {